splitted AsyncFunctionExceptionAssertionSpecs
